Wrote a bunch of test to ensure the attacks were performing correctly. Which they seem to be.

+++ Audiabolikal/src/audiabolikal/attacking/DamageDetail.java
@@ -138,7 +138,7 @@ public class DamageDetail {
 		this.rebuffChance3_ = rebuffChance3;
 	}
 
-	public float getRepositionMove() {
+	public int getRepositionMove() {
 		return repositionMove_;
 	}
